% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions_IFN_model_Vol.R
\name{BD_vol}
\alias{BD_vol}
\title{Prépare une base de données de complétes d'arbres avec une hauteur de décrochement
à partir des données brutes de l'IFN}
\usage{
BD_vol(IFN_data, code_ess)
}
\arguments{
\item{IFN_data}{Base de données de l'inventaire forestier national : liste
comprenant les base de données : arbres (IFNarbres), placettes (IFNplacettes)
et écologie (IFNecologie). Correspond à la base de donnée issue des données brute IFN
et que l'on peut obtenir à partir de la fonction import_IFN du package IFNtraitement}

\item{code_ess}{base de donnée de correspondance entre les codes essences de l'IFN, de l'ONF
et une classification entre familles d'essence (F : feuillus, R : résineux)}
}
\value{
retourne la base de données INF_data rentrée initialement filtrée pour que
la base de donnée arbre (IFNarbres) ne contiennent que des arbres dont on connait à la fois :
l'essence, le diamètre à 1,30m, la hauteur totale, la hauteur de décrochement et la sylvo-ecor-région d'appartenence
}
\description{
Cette fonction prépare une base de donnée issue des données de l'inventaire forestier
qui permetra de calibrer un modèle de prédiction de la hauteur de décrochement à partir
de l'essence, le diamètre, la hauteur total des arbres et leur sylvo-éco-région d'appartenance.
Elle filtre donc la base de donnée initial pour avoir une base de données complète
(i.e. avec pour chaque arbres les 5 données citées précédement)
}
