% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calc_Vol_Emerge.R
\name{Calc_Vol_Emerge}
\alias{Calc_Vol_Emerge}
\title{Calcul un volume à partir des tarifs Emerge en estimant la hauteur de décrochement}
\usage{
Calc_Vol_Emerge(Id_arbre, ess, d13, htot, X = NULL, Y = NULL,
  crs = NA)
}
\arguments{
\item{Id_arbre}{identifiant unique de l'arbre (numerique ou chaine de caractère)}

\item{ess}{code de l'essence selon code ONF (voir colonne Cod_ess de la base de données emergeIFNVol::Code_ess_Emerge)}

\item{d13}{diamètre à hauteur de poitrine de l'arbre}

\item{htot}{hauteur total de l'arbre}

\item{X}{coordonnée X de l'arbre (permet de lui affecter une sylvo-éco-région d'appartenance, ATTENTION au système de coordonnée de référence = crs)}

\item{Y}{coordonnée Y de l'arbre (permet de lui affecter une sylvo-éco-région d'appartenance, ATTENTION au système de coordonnée de référence = crs)}

\item{crs}{système de coordonnée de référence rattaché aux coordonnées X et Y renseigné précédement (par défaut : Lambert 93 code EPSG = 2154)}
}
\description{
Cette fonction permet de calculer un volume en utilisant les tarifs Emerge mais
en ne rensignant "que" l'essence, le diamètre à hauteur de poitrine
et la hauteur d'un ou plusieurs arbres. La hauteur de décrochement est
prédite en utilisant un modèle de prédiction établit à partir des
données des placettes de l'inventaire foretier national
}
