% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emergeIFNVol_plot.R
\name{graph_res_modelisation}
\alias{graph_res_modelisation}
\title{graphiques d'analyse des modélisation de la hauteur de décrochement}
\usage{
graph_res_modelisation(dsn = getwd(), res_modelisation)
}
\arguments{
\item{dsn}{chemin complet vers le dossier du projet comportant les sous-dossiers
"modeles/NumDep_..._NomSER_.../graphs"}

\item{res_modelisation}{liste de résultat issue de la fonction emergeIFNVol::modelisation(...) (voir aide spécifique de cette fonction)}
}
\value{
Cette fonction ne retourne rien mais enregistre des fichiers images
}
\description{
Cette fonction enregistre, pour chaque modèles issues de la fonction emergeIFNVol::modelisation(...),
les graphiques d'analyse de ces modèles avec une image par modèle
}
