% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelisation_hdec.R
\name{modeles_hdec_IFN}
\alias{modeles_hdec_IFN}
\title{Fonction unique interragissant avec l'utilisateur et réalisant l'ensemble des actions liées
à la modélisation de hauteur de décrochement
(de l'import des données brutes IFN à la calibration de modèles et sorties graphiques)}
\usage{
modeles_hdec_IFN(dos_projet = NULL)
}
\arguments{
\item{dos_projet}{chemin vers le dossier où l'on enregistre toutes les fichiers issue de la fonction (par défaut vaut NULL)}
}
\value{
Cette fonction renvoie le chemin vers le dossier où sont enregistré les modéles qui ont
été calibrés ainsi que les graphiques d'analyse de ces modèles. Elle enregistre en fait sous
forme d'archive RDS le'on résultat de la fonction emergeIFNVol::modelisation(...)
}
\description{
Cette fonction permet d'effectuer en une seule étape l'ensemble des modélisations de
prédiction de la hauteur de décrochement. Cette fonction interagie avec l'utilisateur et fait appel
aux fonctions suivantes : IFNtraitement::import_IFN(...), IFNtraitement::classif_cat_Diam(...)
emergeIFNVol::BD_Vol(...), emergeIFNVol::Nb_arbre_hdec_bois_fort(...),
emergeIFNVol::modelisation(...), emergeIFNVol::graph_res_modelisation(...)
}
