% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions_IFN_model_Vol.R
\name{modelisation}
\alias{modelisation}
\title{Calibration de modèles de prédiction de la hauteur de décrochement d'arbres à partir de
la base de donnée de l'inventaire forestier nationale (=IFN)}
\usage{
modelisation(IFN_data, tbl_corresp_model)
}
\arguments{
\item{IFN_data}{Base de données de l'inventaire forestier national préalablement filtrée et
issue de la fonction emergeIFNVol::BD_Vol(...) (voir l'aide spécifique à cette fonction pour plus d'informations)}

\item{tbl_corresp_model}{tableau de correspondance définissant si pour une essence et une
sylvo-éco-région (=SER) donnée on effectuera une modélisation spécifique à ce couple essence/SER
tableau de correspondance issue de la fonction emergeIFNVol::Nb_arbre_hdec_bois_fort(...) (voir aide spécifique à cette fonction pour plus d'informations)}
}
\value{
renvoie une liste comportant une sous-liste par sylvo-écorégion (SER)
et une sous-liste toutes SER confondues. ces sous listes comportant autant de sous-liste que
de modèles spécifique à une essence + deux sous listes correspondant
aux  modèles résineux et feuillus. Les listes par modèles comportent elles :
la base de données arbres ayant servis à la modélisation, la partie de cette base de données ayant servie
à la calbration, celles ayant servie à la validation du modèle, le modèle en lui-même et
les graphiques d'analyse du modèle
}
\description{
Calibration de modèles de prédiction de la hauteur de décrochement d'arbres à partir de
la base de donnée de l'inventaire forestier nationale (=IFN)
}
