% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emergeIFNVol_plot.R
\name{multiplot}
\alias{multiplot}
\title{fonction assurant la mise en page d'une liste de graphique à enregister sous une unique image}
\usage{
multiplot(plots, file = ".", width = 1000, height = 1000, cols = 1,
  layout = NULL)
}
\arguments{
\item{plots}{liste des graphiques à enregistrer dans une seule image}

\item{file}{chemin complet de l'image qui sera enregistrer en sortie
de la fonction (ex : "c:/dos1/dos2/graph1.png")}

\item{width}{largeur de l'image (par défaut = 1000)}

\item{height}{hauteur de l'image (par défaut = 1000)}

\item{cols}{nombre de colonnes selon lesquels les graphiques devront être
réparties dans l'image final (par défaut vaut 1, donnée non prise en compte
si l'argument layout est renseigné)}

\item{layout}{matrice permettant de spécifier l'agencement des différents graphiques
(ex : si on 3 graphiques spécifié dans plots et layout = matrix(c(1,1,3,2), nrow=2, byrow=TRUE)
alors l'image final sera composée d'une première ligne avec le premier graphique de plots
puis d'une seconde ligne divisée en 2 parties avec à gauche le 3ieme graphique de plots
et à droite le 2nd graphiques de plots)}
}
\description{
fonction assurant la mise en page d'une liste de graphique à enregister sous une unique image
}
