% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions_IFN_model_Vol.R
\name{Nb_arbre_hdec_bois_fort}
\alias{Nb_arbre_hdec_bois_fort}
\title{Détermine par couple essence/sylvo-éco-région si une modélisation spécifique au couple est possible
(nombre de données disponible > seuil minimal d'abres)}
\usage{
Nb_arbre_hdec_bois_fort(IFN_vol, code_ess, nb_arbre_min = 200)
}
\arguments{
\item{IFN_vol}{Base de données de l'inventaire forestier national préalablement filtrée et
issue de la fonction emergeIFNVol::BD_Vol(...) (voir l'aide spécifique à cette fonction pour plus d'informations)}

\item{code_ess}{base de donnée de correspondance entre les codes essences de l'IFN, de l'ONF
et une classification entre familles d'essence (F : feuillus, R : résineux)}

\item{nb_arbre_min}{nombre d'arbre minimum nécessaire à une modélisation de la hauteur de décrochement
spécifique à un couple essence/sylvo-éco-région}
}
\value{
tableau de correspondance définissant si pour une essence et une
sylvo-éco-région ( = SER) donnée on possède un nombre suffisant de données (d'arbres) pour effectuer
une modélisation spécifique à ce couple essence/SER. ou à défaut si on effectue une modélisation différenciant
feuillus et résineux par sylvo-éco-région (=SER) ou toutes SER confondues
}
\description{
A partir de la base de données de l'inventaire forestier national filtrée par la fonction
emergeIFNVol::BD_Vol(...), cette fonction prépare la modélisation de la hauteur de décrochement
en retournant un tableau de correspondance définissant si pour une essence et une
sylvo-éco-région ( = SER) donnée on possède un nombre suffisant de données (d'arbres) pour effectuer
une modélisation spécifique à ce couple essence/SER. Si ce n'est pas le cas on spécifie si on effectuera
une modélisation spécifique à la SER en distinguant feuillus et résineux ou si on effectuera une
modélisation tout SER confondu en distinguant feuillus et résineux
}
