% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions_IFN_model_Vol.R
\name{predict_hdec}
\alias{predict_hdec}
\title{Prédiction de la hauteur de décrochement d'arbres selon leur essence, diamètre, hauteur total et sylvo-éco-région}
\usage{
predict_hdec(res_modelisation, df)
}
\arguments{
\item{res_modelisation}{liste de résultat issue de la fonction emergeIFNVol::modelisation(...) (voir aide spécifique de cette fonction)}

\item{df}{data.frame des données par arbres et pour lesquels on veut prédir
la hauteur de décrochement. Ce data.frame doit comporter pour chaque arbre
les données suivantes :ida, ser, code, fam, htot, c13}
}
\value{
renvoir le data.frame df initial auquel s'ajoute la colonne hdec correspondant à a hauteur de décrochement prédite
}
\description{
Cette fonction à partir d'une base de données arbres pour lesquels sont spécifié l'essence,
le diamètre à 1,30m, la hauteur total et la sylvo-éco-région d'appartenance prédit
la hauteur de décrocement de ces arbres à partir des modèles de prédictions issue de la fonction emergeIFNVol::modelisation(...)
}
